% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_divider.R
\name{xxx_divider1}
\alias{xxx_divider1}
\alias{xxx_divider2}
\title{Insert a divider (line break or page break)}
\usage{
xxx_divider1(mystring = NULL)

xxx_divider2(mystring = NULL)
}
\arguments{
\item{mystring}{the content of your break, must be one line only, not too long}
}
\value{
A character string vector returned invisibly,
but automatically displayed in the console, the output will be saved into clipboard
}
\description{
For user guide, please refer to \url{https://github.com/Hzhang-ouce/ARTofR}\cr
}
\details{
Divider is shorter than title for hierarchy purpose, title will be come a section (for folding) in Rstudio, but divider will not\cr
There are two styles of break, xxx_divider1 and xxx_divider2 \cr
Please note that a super long divider string will result in an error\cr
call \code{xxx_divider1()} will grab text from clipboard,\cr
call \code{xxx_divider1('')} will make a line without text, \cr
call \code{xxx_divider1('mystring')} will make a divider with 'mystring'
}
\section{Functions}{
\itemize{
\item \code{xxx_divider2}: Same as xxx_divider1 but different style
}}

\examples{
# COMMON WORKFLOW OF THIS PACKAGE
# 1. type your text in any text editor
# 2. copy your text
# 3. call xxx_box() etc without arguments to grab text from clipboard
# 4. decorated text will be displayed and sent to clipboard, NO NEED to copy them
# 5. paste into your R script
#
#
#...............................................
# # OTHER USAGE

xxx_divider1('mystring')
xxx_divider1('')
xxx_divider2('mystring')
xxx_divider2('')
}
