% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_divider.R
\name{xxx_divider2}
\alias{xxx_divider2}
\alias{xxx_divider1}
\title{Insert a divider (line break or page break)}
\usage{
xxx_divider2(mystring = NULL)

xxx_divider1(mystring = NULL)
}
\arguments{
\item{mystring}{the content of your break, must be one line only, not too long}
}
\value{
A character string vector returned invisibly,
but automatically displayed in the console, the output will be saved into clipboard
}
\description{
Please refer to the github page <github.com/Hzhang-ouce/ARTofR> for more instructions.\cr
}
\details{
Divider is shorter than title for hierarchy purpose\cr
There are two styles of break, xxx_divider1 and xxx_divider2 \cr

call \code{xxx_divider1()} will grab text from clipboard,\cr
call \code{xxx_divider1('')} will make a line without text, \cr
call \code{xxx_divider1('mystring')} will make a divider with 'mystring'
}
\section{Functions}{
\itemize{
\item \code{xxx_divider1}: Same as xxx_divider2 but different style
}}

\examples{
#' # COMMON WORKFLOW OF THIS PACKAGE
# 1. type your text in any text editor
# 2. copy your text
# 3. call xxx_box() or xxx_box1() etc
# 4. Check the output in console, NO NEED to copy them
# 4. paste into your R script
#
#
#' # COMMON USAGE
#  xxx_divider1()
#  This will grab text from clipboard
#  xxx_divider1('')
#...........................................
#  xxx_divider1('XXXXX')
#...............XXXXX.......................
# pls note that you must provide a sentence less than 65 character
# call xxx_divider1() will grab text from clipboard,
# call xxx_divider1('') will make a line without text,
# call xxx_divider1('mystring') will make a divider with 'mystring'
}
