% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_title.R
\name{xxx_title1}
\alias{xxx_title1}
\alias{xxx_title2}
\alias{xxx_title3}
\title{Insert a title}
\usage{
xxx_title1(mystring = NULL)

xxx_title2(mystring = NULL)

xxx_title3(mystring = NULL)
}
\arguments{
\item{mystring}{the content of your title, must be one line only, not too long}
}
\value{
A character string vector returned invisibly,
but automatically displayed in the console, the output will be saved into clipboard
}
\description{
Please refer to the github page <github.com/Hzhang-ouce/ARTofR> for more instructions.\cr
}
\details{
There are three levels of titles, inserting a title with xxx_title() will also add a section in Rstudio automatically.\cr
xxx_title1() is the largest title.\cr
xxx_title2() is the middle size title.\cr
xxx_title3() is the smallest title.\cr
}
\section{Functions}{
\itemize{
\item \code{xxx_title2}: the middle size title

\item \code{xxx_title3}: the smallest title
}}

\examples{
#' # COMMON WORKFLOW OF THIS PACKAGE
# 1. type your text in any text editor
# 2. copy your text
# 3. call xxx_box() or xxx_box1() etc
# 4. Check the output in console, NO NEED to copy them
# 4. paste into your R script

#...............................................
#' # COMMON USAGE
#  xxx_title1()
#  xxx_title1('XXXXX')
# pls note that a long title string will make output ugly
}
